feat: 세션 가져오기 성공
