+++ .vscode/settings.json
@@ -1,0 +1,3 @@
+{
+  "compile-hero.disable-compile-files-on-did-save-code": true
+}
