+++ src/bounded-context/user/constants/query-key/ReactQueryKey.ts
@@ -1,0 +1,13 @@
+
+
+
+const user = {
+  session: ["session"]
+}
+
+
+
+
+export const QueryKeys = {
+  user
+}
