--- src/constants/response-code/ResponseCode.ts
@@ -1,19 +1,0 @@
-
-
-
-// ======[ common ]======
-const SUCCESS = "SUCCESS";
-const common = {
-  SUCCESS
-};
-
-// ======[ clientOnly ]======
-const LOADING = "LOADING";
-const clientOnly = {
-  LOADING
-};
-
-export const CodeType = {
-  common,
-  clientOnly
-}
