+++ src/global/config/next-dynamic-config.ts
@@ -1,0 +1,6 @@
+import dynamic from 'next/dynamic';
+
+export const RootAuthFromLocalStorage = dynamic(
+  () => import('@/global/hook/useRootAuth').then(mod => mod.useRootAuth), 
+  { ssr: false }
+);
