--- src/types/common.ts
@@ -1,5 +1,0 @@
-interface GlobalResponse<T> {
-  code: string;
-  data: T;
-  message: string;
-}
